+++ README.md
@@ -9,9 +9,14 @@ install it yourself as:
     $ gem install csa
 
 ## Usage
-
-csa [Project Name][template url]
-
+csa [Project Name] [default template url], or
+```shell
+csa [options]
+    -n, --name NAME                  The Name of the project
+    -d, --dir DIR                    The DIR of the template
+    -u, --url URL                    The URL of the template
+    -h, --help                       Prints help
+```
 ## Development
 
 After checking out the repo, run `bin/setup` to install dependencies. You can also run `bin/console` for an interactive prompt that will allow you to experiment.
